+++ subterra-runtime/src/main/java/io/toterra/subterra/runtime/worldgen/gen/WorldgenConfig.java
@@ -1,4 +1,4 @@
-package io.toterra.subterra.worldgen.gen;
+package io.toterra.subterra.runtime.worldgen.gen;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;